Additional features for subsequent versions
While the initial features are good, I propose a few additions now so they can be considered and during the development stage.
My focus is on quick screen tutorials for clients who need help with infrequently used options (change a setting, add an email address, etc).  A GIF is perfect for this.
Protecting ancillary information is needed (I may have another client's file in a Recently Opened list), so I propose a blur/redaction feature.
Having recipients eyes go to the right place is a priority, so I propose adding Callouts with movable arrows to the Add Text feature. 
Size of resulting GIF is important, and doesn't not necessarily need to be as high as the original screen resolution.  Position or motion may be more important than reading the text of a file menu.
I also propose being able to trim sections inside the video (this may have been implied by the original spec, but I'm making it explicit).  It sometimes takes a few seconds of mouse hunting to find the next object to click; the user doesn't need to see all the hunting, or it could be sped up faster than real-time.  Otherwise the 6.00 second maximum time may have to be increased for mini-tutorial purposes.